Refactor code in main.rb
